Separate LED and PST activation using separate pins.
The primary problem is that when powered via USB connected to a computer, the
Trinket detects the computer and enters a 12 second bootloader loop, waiting
for new firmware. During this interval, the led is blinked twice a second...
not good for the PowerSwitch.

The update dual-updates the LED pin (#1) and the pin for the PST (#0) in tandem
during runtime operation.